requirements.txt
Flask==3.1.0
flask_cors==5.0.1
numpy==2.2.4
opencv_python==4.11.0.86
Pillow==11.2.0
Requests==2.32.3
scikit_learn==1.6.1
sentence_transformers==4.0.1
torch==2.6.0
transformers==4.49.0
